fix: [Multi-Factor Administrator] group suggester not working - EXO-67772 (#55)
Before this change, when you open the group management drawer in protected access to the platform, the group suggester is not displayed and cannot delete old saved groups. To resolve this issue, replace exo-group-suggester with exo-identity-suggester. After this change, the group suggester displays spaces and is able to remove old groups saved in the list.